requirements.txt
dicttoxml==1.5.5
diff-match-patch==20121119
Django==1.8.2
django-braces==1.4.0
django-crispy-forms==1.4.0
django-extensions==1.2.5
django-filter==0.10.0
django-import-export==0.2.3
django-multiupload==0.3
django-recaptcha==1.0.3
django-tables2==1.0.4
djangorestframework==3.1.3
gdata==2.0.18
google-api-python-client==1.2
httplib2==0.9
image==1.3.2
mongoengine==0.9.0
MySQL-python==1.2.5
Pillow==2.5.1
Pygments==1.6
pymongo==2.8
python-ldap==2.4.15
requests==2.4.1
selenium==2.42.1
six==1.9.0
tablib==0.10.0
vboxapi==1.0
xlrd==0.9.3
